Improved plug-and-play DX, hardened AOT, and removed sensitive-data attribute
  - Added `ICqrsDispatcher` abstraction (Send<T>, Send(object), Publish<T>) and `ExecutionScopeMode` (Current/New)
  - Added notification pipeline behaviors and request exception hooks (actions/handlers) with source-gen invokers (no
  reflection)
  - Added untyped/dynamic request execution API
  - Hardened core behaviors: conditional rate limiting + strong IDs, dispose linked CTS in timeout, resilience tweaks
  - Improved generators: root AOT hints via module initializer, add handler selection diagnostics + config knobs,
  reorganize generator layout
  - Added `AddCqrsPipelinePack()` one-call registration and expand Sample self-test to exercise major features (incl. AOT
  publish run)
  - Packaged `CQRSharp.Generators` as analyzer-only and keep meta-package embedding under `analyzers/`
  - Removed `[SensitiveData]` and related metadata plumbing; add/update unit tests

+++ src/Application/CQRSharp.Pipelines/Options/RateLimiterOptions.cs
@@ -3,7 +3,7 @@
 namespace CQRSharp.Pipelines.Options;
 
 /// <summary>
-///     Class object for the options needed for the rate limited.
+///     Options for configuring the built-in <see cref="RateLimiter" />.
 /// </summary>
 public sealed class RateLimiterOptions
 {
@@ -15,8 +15,7 @@ public sealed class RateLimiterOptions
     /// <summary>
     ///     How many tokens are to be replenished for each user per second.
     /// </summary>
-    //TODO: This is atrocious and the customization is quite bad.
-    public int ReplenishRatePerSecond { get; set; } = 1;
+    public double ReplenishRatePerSecond { get; set; } = 1;
 
     /// <summary>
     ///     The scope of the rate limiter.
